6.FastAPI/app/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
google-cloud-storage==2.11.0
torch==2.1.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.26.1
pandas==2.1.1
scikit-learn==1.3.1
xgboost==2.0.0
catboost==1.2.2
matplotlib==3.8.0
python-multipart==0.0.6